requirements.txt
anytree
aria2p
asyncio
beautifulsoup4
bencoding
cfscrape
dnspython
feedparser
flask
flask-restful
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
git+https://github.com/zevtyardt/lk21.git
lxml
pillow
psutil
pybase64
pymongo
pyrogram
python-dotenv
python-magic
python-telegram-bot==13.15
qbittorrent-api
requests
telegraph
tenacity
tgCrypto
xattr
yt-dlp
